requirements.txt
# TikTok Transcription Pipeline - Requirements
# Core dependencies for production transcription pipeline

# Video/Audio Processing
yt-dlp>=2023.12.30
faster-whisper>=0.10.0
ffmpeg-python>=0.2.0

# Data Processing
pandas>=2.1.4
numpy>=1.24.4

# Utilities
python-dotenv>=1.0.0
pathlib>=1.0.1

# Optional: Progress bars (recommended)
tqdm>=4.66.0

# Topic Extraction & NLP (Step 2)
keybert>=0.8.4
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
nltk>=3.8.1

# API Server (Step 2)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Semantic Search (Step 3)
faiss-cpu>=1.7.4
